+ save competitions

+++ src/main/java/pro/metrus/fvsr/repositories/VidRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import pro.metrus.fvsr.domains.Vid;
 
 public interface VidRepository
-    extends JpaRepository<Vid, Long> {
+    extends JpaRepository<Vid, Short> {
 }
